! r upd json view
