+++ packages/ispect/lib/src/common/extensions/history.dart
@@ -28,7 +28,7 @@ extension HistoryListFlutterText on List<ISpectifyData> {
     final sb = StringBuffer();
     for (final data in this) {
       sb
-        ..writeln(data.textMessage) // Ensures newline after text.
+        ..writeln(data.textMessage.truncated) // Ensures newline after text.
         ..writeln(ConsoleUtils.bottomLine(30)); // Separator line.
     }
     return sb.toString().trim(); // Trim trailing newline for cleaner output.
