+++ packages/ispect/lib/src/features/json_viewer/pretty_json_view.dart
@@ -1,4 +1,2 @@
-export 'src/json_tree_node.dart';
-export 'src/json_tree_view.dart';
 export 'src/models/json_node.dart';
 export 'src/theme/json_view_theme.dart';
